# api/openapi.yaml
openapi: '3.0.3'
info:
  description: CSE138 Assignment 2
  version: 1.0.0
  title: CSE138 Assignment 2
paths:
  /ping:
    get:
      description: Check if node is initialized
      responses:
        200:
          description: OK
  /data/{key}:
    put:
      description: Creates/updates a key-value pair
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value:
                  type: string
              required:
                - value
              additionalProperties: false
      responses:
        200:
          description: Key-value pair updated
        201:
          description: Key-value pair created
    get:
      description: Gets value of key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Value of key
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    type: string
                required:
                  - value
                additionalProperties: false
        404:
          description: Key not found
    delete:
      description: Deletes key from store
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Key deleted
        404:
          description: Key not found
  /data:
    get:
      description: Get all key-value pairs
      responses:
        200:
          description: A JSON object of key-value pairs
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
  /view:
    put:
      description: Updates view
      requestBody:
        description: Request Body
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                view:
                  type: array
                  items:
                    type: object
                    properties:
                      address:
                        type: string
                      id:
                        type: number
                    required:
                      - address
                      - id
                    additionalProperties: false
              required:
                - view
              additionalProperties: false
      responses:
        200:
          description: View updated 